legacy/LinkedInJobsScraper/requirements.txt
python-dotenv
droidrun[google,anthropic,openai,deepseek,ollama,dev]
llama-index-llms-google-genai
colorama